requirements.txt
# Core dependencies
playwright==1.40.0
supabase==2.3.0
apscheduler==3.10.4
pandas==2.1.4
python-dotenv==1.0.0
fake-useragent==1.4.0

# Data processing
beautifulsoup4==4.12.2
lxml==5.1.0
python-dateutil==2.8.2
trafilatura>=1.6.0
newspaper3k>=0.2.8

# HTTP and networking
httpx>=0.24.0,<0.25.0
aiohttp==3.9.1

# Logging and utilities
python-json-logger==2.0.7

# Email notifications (optional)
aiosmtplib==3.0.1

# Testing (optional, for development)
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-mock==3.12.0
